+++ caffe2/CMakeLists.txt
@@ -16,7 +16,6 @@ add_subdirectory(distributed)
 add_subdirectory(image)
 add_subdirectory(video)
 add_subdirectory(mkl)
-add_subdirectory(ideep)
 add_subdirectory(mobile)
 add_subdirectory(mpi)
 add_subdirectory(observers)
