+++ caffe2/contrib/ideep/operators/dropout_op.cc
@@ -1,4 +1,4 @@
-#include <caffe2/ideep/ideep_utils.h>
+#include <caffe2/contrib/ideep/ideep_utils.h>
 
 namespace caffe2 {
