+++ caffe2/contrib/ideep/operators/operator_fallback_ideep.h
@@ -3,7 +3,7 @@
 #include <caffe2/core/common.h>
 #include <caffe2/core/context.h>
 #include <caffe2/core/operator.h>
-#include <caffe2/ideep/ideep_utils.h>
+#include <caffe2/contrib/ideep/ideep_utils.h>
 #include <caffe2/proto/caffe2.pb.h>
 
 namespace caffe2 {
