+++ cmake/Summary.cmake
@@ -70,6 +70,10 @@ function (caffe2_print_configuration_summary)
   endif()
   message(STATUS "  USE_METAL             : ${USE_METAL}")
   message(STATUS "  USE_MKL               : ${USE_MKL}")
+  if(${USE_MKL})
+    message(STATUS "    MKL_USE_MKLML       : ${MKL_USE_MKLML}")
+    message(STATUS "    MKL_USE_IDEEP       : ${MKL_USE_IDEEP}")
+  endif()
   message(STATUS "  USE_MOBILE_OPENGL     : ${USE_MOBILE_OPENGL}")
   message(STATUS "  USE_MPI               : ${USE_MPI}")
   message(STATUS "  USE_NCCL              : ${USE_NCCL}")
